tests/wpt/tests/speculation-rules/prefetch/different-initiators-2.https.html

<!DOCTYPE html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.sub.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>
setup(() => assertSpeculationRulesIsSupported());

// Regression test for https://crbug.com/1431804.
promise_test(async t => {
  const win = await spawnWindow(t, { protocol: 'https' });
  const nextUrl = win.getExecutorURL({ protocol: 'https', page: 2 });

  // Navigate `win` from Document #1 -> #2 (nextUrl) -> #3 (tempUrl) ->
  // #4 (nextUrl),
  // Start speculation rules prefetch from #1, and
  // Try using the prefetched result for the navigation #3 -> #4.
  // The Documents #2 and #4 are different, but the same RenderFrameHost is
  // used before https://crbug.com/936696 is done.

  await win.forceSinglePrefetch(nextUrl);

  // Register a SW for `nextUrl` -- this is a trick to make the prefetched
  // result to put in `PrefetchService::prefetches_ready_to_serve_` in
  // Chromium implementation but actually not used by this navigation.
  const r = await service_worker_unregister_and_register(
      t, 'resources/sw.js', nextUrl);
  await wait_for_state(t, r.installing, 'activated');

  // Navigate #1 -> #2.
  // This doesn't use the prefetched result due to the ServiceWorker.
  await win.navigate(nextUrl);

  // Unregister the SW.
  await service_worker_unregister(t, nextUrl);

  // Navigate #2 -> #3 -> #4.
  const tempUrl = win.getExecutorURL({ protocol: 'https', page: 3 });
  await win.navigate(tempUrl);
  await win.navigate(nextUrl);

  const headers = await win.execute_script(() => {
    return requestHeaders;
  }, []);
  assert_not_prefetched(headers,
      "Prefetch should not work for different initiators.");
}, "Prefetches from different initiator Documents with same RenderFrameHost");
</script>
