third_party/blink/web_tests/virtual/prefetch/external/wpt/speculation-rules/prefetch/different-initiators-2.https-expected.txt

This is a testharness.js-based test.
FAIL Prefetches from different initiator Documents with same RenderFrameHost assert_equals: Prefetch should not work for different initiators. expected "" but got "prefetch"
Harness: the test ran to completion.
